Add ecm_add_app_icon function.
This adds an application icon to an executable from PNG files for
Windows and Mac OS X. Unlike the similar kde4_add_app_icon macro from
kdelibs, this requires icons to be explicitly listed as arguments
(meaning CMake can tell when ones are added or deleted and reconfigure
as appropriate), and it works with Matthias Benkmann's png2ico tool, as
well as the KDE-Win tool of the same name.

Currently missing unit tests. Also completely untested (except that
`make test` runs on Linux, so there are no obvious syntax errors).

With thanks to Ralf Habacker for the inital work on porting
kde4_add_app_icon.

CHANGELOG: Add ECMAddAppIcon module to add icons to executable targets
on Windows and Mac OS X.