Fix FindEGL compile check when using Emscripten
Under Emscripten, EGL does not require any libraries to be linked in.
The FindEGL module already has a partial concession of this since
4faa4fd271b9ac3f13a01e1fd02bef9cecee8c2f, but it still searches for the
library and, crucially, tries to link it when performing the compile
check if found. If the system you're on has a libEGL.so installed, this
will end up trying to link that into the build, which will obviously
fail because you can't like a native shared object into a WebAssembly
build.

This commit removes the library part entirely under Emscripten, making
the find module no longer go searching for it pointlessly and not
attempt to link anything when performing the compile check.